two colliding files
Actual SHA1 collision in the wild due to
https://shattered.io/static/shattered.pdf

good.pdf and bad.pdf have different content, but the same SHA1

Why does git generate two different SHAs for the blobs for
these two files? It's because git prefixes the type and length,
and this SHA1 collision does not work if additional data is prefixed to
the file. Although suffixing added data will result in other files with
the same SHA.

PS, see the previous commit in this repo for another interesting
problem.